# server/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SocialFlow AI - Phase 1 API Specs",
    "version": "1.0.0",
    "description": "Production-grade endpoints for Authentication, Workspace collaboration, and RBAC members administration."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/workspace": {
      "post": {
        "summary": "Create a new organizational workspace",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 2, "maxLength": 50 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Workspace created successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/posts": {
      "post": {
        "summary": "Create a new post (draft or scheduled)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content", "platforms"],
                "properties": {
                  "content": { "type": "string" },
                  "platforms": {
                    "type": "array",
                    "items": { "type": "string", "enum": ["twitter", "instagram", "facebook", "linkedin", "youtube", "threads"] }
                  },
                  "status": { "type": "string", "enum": ["draft", "scheduled"], "default": "draft" },
                  "scheduledAt": { "type": "string", "format": "date-time" },
                  "media": { "type": "array", "items": { "type": "string" } },
                  "platformContent": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Post successfully created" },
          "400": { "description": "Validation failed / schedule constraints violated" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/comments": {
      "get": {
        "summary": "Retrieve comments for connected accounts in the workspace",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "platform",
            "in": "query",
            "schema": { "type": "string", "enum": ["twitter", "instagram", "facebook", "linkedin", "youtube", "threads"] }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["unresolved", "resolved"] }
          },
          {
            "name": "assignedTo",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Comments list retrieved" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden workspace access" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}